chore(!TE-75) : Refactored the Toaster into seperate component
